+++ lib/danger/request_sources/gitlab.rb
@@ -65,12 +65,7 @@ module Danger
       end
 
       def base_commit
-        if self.commits_json.last
-          first_commit_in_branch = self.commits_json.last.id
-          @base_commit ||= self.scm.exec "rev-parse #{first_commit_in_branch}^1"
-        else
-          @base_commit = ""
-        end
+        @base_commit ||= self.mr_json.diff_refs.base_sha
       end
 
       def mr_comments
@@ -89,10 +84,11 @@ module Danger
       end
 
       def setup_danger_branches
-        raise "Are you running `danger local/pr` against the correct repository? Also this can happen if you run danger on MR without changes" if base_commit.empty?
         base_branch = self.mr_json.source_branch
         head_branch = self.mr_json.target_branch
         head_commit = self.scm.head_commit
+
+        raise "Are you running `danger local/pr` against the correct repository? Also this can happen if you run danger on MR without changes" if base_commit == head_commit
 
         # Next, we want to ensure that we have a version of the current branch at a known location
         scm.ensure_commitish_exists_on_branch! base_branch, base_commit
@@ -106,9 +102,6 @@ module Danger
 
       def fetch_details
         self.mr_json = client.merge_request(ci_source.repo_slug, self.ci_source.pull_request_id)
-        self.commits_json = client.merge_request_commits(
-          ci_source.repo_slug, self.ci_source.pull_request_id
-        ).auto_paginate
         self.ignored_violations = ignored_violations_from_pr
       end