+++ spec/lib/danger/request_sources/gitlab_spec.rb
@@ -105,11 +105,6 @@ RSpec.describe Danger::RequestSources::GitLab, host: :gitlab do
         "k0nserv%2Fdanger-test",
         593_728
       )
-      stub_merge_request_commits(
-        "merge_request_593728_commits_response",
-        "k0nserv%2Fdanger-test",
-        593_728
-      )
       @comments_stub = stub_merge_request_comments(
         "merge_request_593728_comments_response",
         "k0nserv%2Fdanger-test",
@@ -120,33 +115,18 @@ RSpec.describe Danger::RequestSources::GitLab, host: :gitlab do
     it "determines the correct base_commit" do
       subject.fetch_details
 
-      expect(subject.scm).to receive(:exec)
-        .with("rev-parse adae7c389e1d261da744565fdd5fdebf16e559d1^1")
-        .and_return("0e4db308b6579f7cc733e5a354e026b272e1c076")
-
       expect(subject.base_commit).to eq("0e4db308b6579f7cc733e5a354e026b272e1c076")
     end
 
-    context "works also on empty MR" do
-      before do
-        stub_merge_request_commits(
-          "merge_request_593728_no_commits_response",
-          "k0nserv%2Fdanger-test",
-          593_728
-        )
-      end
+    it "raise error on empty MR" do
+      subject.fetch_details
 
-      it "return empty string if last commit do not exist" do
-        subject.fetch_details
+      expect(subject.scm).to receive(:head_commit).
+        and_return("345e74fabb2fecea93091e8925b1a7a208b48ba6")
+      expect(subject).to receive(:base_commit).
+        and_return("345e74fabb2fecea93091e8925b1a7a208b48ba6")
 
-        expect(subject.base_commit).to eq("")
-      end
-
-      it "raise error on empty commit" do
-        subject.fetch_details
-
-        expect { subject.setup_danger_branches }.to raise_error("Are you running `danger local/pr` against the correct repository? Also this can happen if you run danger on MR without changes")
-      end
+      expect { subject.setup_danger_branches }.to raise_error("Are you running `danger local/pr` against the correct repository? Also this can happen if you run danger on MR without changes")
     end
 
     it "setups the danger branches" do
@@ -174,12 +154,6 @@ RSpec.describe Danger::RequestSources::GitLab, host: :gitlab do
       subject.fetch_details
 
       expect(subject.mr_json).to be_truthy
-    end
-
-    it "sets its commits_json" do
-      subject.fetch_details
-
-      expect(subject.commits_json).to be_truthy
     end
 
     it "sets its ignored_violations_from_pr" do