+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ## master
 * Fixes importing a Dangerfile from a GitLab repository [@koffeinfrei](https://github.com/koffeinfrei/)
 * Adds `dry_run` command to allow running danger on localhost without actual PR/MR [@otaznik-net](https://github.com/otaznik-net)
+* GitLab: Determine base_commit from the MR API instead of querying the commits [@leipert](https://github.com/leipert)
 
 ## 5.6.7
 
@@ -35,7 +36,7 @@
 
 ## 5.6.2
 
-* Update Screwdriver CI to parse repo_slug/repo_url correctly. [@fandyfyf](https://github.com/fandyfyf) 
+* Update Screwdriver CI to parse repo_slug/repo_url correctly. [@fandyfyf](https://github.com/fandyfyf)
 
 ## 5.6.1